+++ package.json
@@ -28,6 +28,7 @@
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-navigation-menu": "^1.2.13",
     "@radix-ui/react-popover": "^1.1.14",
+    "@radix-ui/react-progress": "^1.1.7",
     "@radix-ui/react-radio-group": "^1.3.7",
     "@radix-ui/react-scroll-area": "^1.2.9",
     "@radix-ui/react-select": "^2.2.5",
@@ -47,7 +48,7 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "framer-motion": "^12.18.1",
-    "lucide-react": "^0.518.0",
+    "lucide-react": "^0.525.0",
     "marked": "^15.0.12",
     "next": "15.3.3",
     "next-themes": "^0.4.6",
@@ -58,6 +59,7 @@
     "react-hook-form": "^7.58.1",
     "react-markdown": "^10.1.0",
     "react-textarea-autosize": "^8.5.9",
+    "recharts": "2.15.4",
     "rehype-external-links": "^3.0.0",
     "rehype-katex": "^6.0.0",
     "remark-gfm": "^3.0.1",
